+++ src/factory/YieldValueProviderFactory.sol
@@ -36,6 +36,10 @@ contract YieldValueProviderFactory is IYieldValueProviderFactory {
             timeScale_
         );
 
+        yieldValueProvider.allowCaller(
+            yieldValueProvider.ANY_SIG(),
+            msg.sender
+        );
         return address(yieldValueProvider);
     }
 }